Compiler portability changes in model folder: 1. throw () specifications are macroed out (via THROW_SPEC     and THROW_SPEC_2 macros) in MSVC builds.  Those     macros are defined in modelmarkov.h.  The Microsoft     compiler warns about every throw specification.     (And a good thing too, throw specifications blow.       It's the compiler's job to nut that stuff out!       C++ isn't Java! C++ has compilers that *work*      for a living). 2. list of mixin models is now loaded via  a function,     loadBuiltInMixInModels, rather than a declaration of     builtin_mixmodels_definition (MSVC complains string     too long).  The function's implemented in     modelmixture.cpp. 3. list of protein models, likewise, via function     loadBuiltInProteinModels, rather than a declaration of     builtin_prot_models.  The function's implemented in     modelprotein.cpp. 4. ModelFactoryMixlen::sortClassesByTreeLength     now uses boost::scoped_array, in MSVC builds, rather     than relying on non-standard dynamic-sized array     on the stack.  And also     (a) ModelMarkov::init_state_freq     (b) ModelMarkov::computeTransMatrix     (c) ModelMarkov::aTimesDiagonalBTimesTransposeOfC     (d) ModelMarkov::computeTransDerv     (e) ModelMarkov::readParameters     (f) ModelMixture::getStateFrequency     (g) ModelPoMo::estimateEmpiricalBoundaryStateFreqs     (h) ModelPoMo::estimateEmpiricalBoundaryStateFreqs     (i) ModelPoMoMixture::decomposeRateMatrix     (j) ModelSet::decomposeRateMatrix     (k) ModelSubst::multiplyWithInvEigenvector     (l) PartitionModel::PartitionModel 5. Declared members of a lot of rate classes (that could     be declared const) const:     (a) RateFree::meanRates() and isGammaRate()     (b) RateGamma::isGammaRate(), getGammaShape(),          getNRate(), getNDiscreteRate().     (c) RateGammaInvar::meanRates().     (d) RateHeterogeneity::getNRate(),          getNDiscreteRate(), getPInvar(), getGammaShape()          isGammaRate() and meanRates().     (e) RateHeterotachy::getNRate() and  getNDiscreteRate().     (f) RateInvar::getPInvar()     (g) RateKategory::getNRate() and getNDiscreteRate().     (h) RateMeyerDiscrete::getNDiscreteRate() 6. RateMeyerHaeseler::computeFuncDerv does its    #pragma omp parallel reductions... on local variables     (because Microsoft's Open MP support refuses to     compile if references to variables are used).

+++ model/modelset.h
@@ -40,7 +40,7 @@ public:
 	 * get the size of transition matrix, default is num_states*num_states.
 	 * can be changed for e.g. site-specific model
 	 */
-	virtual int getTransMatrixSize() { return num_states * num_states * size(); }
+	virtual int getTransMatrixSize() { return num_states * num_states * static_cast<int>(size()); }
 
 
 	/**
