+++ model/rategamma.cpp
@@ -259,12 +259,12 @@ int RateGamma::computePatternRates(DoubleVector &pattern_rates, IntVector &patte
 
 	phylo_tree->computePatternLhCat(WSL_RATECAT);
 
-	int npattern = phylo_tree->aln->getNPattern();
+	intptr_t npattern = phylo_tree->aln->getNPattern();
 	pattern_rates.resize(npattern);
 	pattern_cat.resize(npattern);
 
     double *lh_cat = phylo_tree->tree_buffers._pattern_lh_cat;
-	for (int i = 0; i < npattern; i++) {
+	for (intptr_t i = 0; i < npattern; i++) {
 		double sum_rate = 0.0, sum_lh = 0.0;
 		int best = 0;
 		for (int c = 0; c < ncategory; c++) {
